Rename rusoto_extra.rs to post_policy.rs
